Add `trigger` in event `detail`
`detail.elt` can sometimes be the element that triggered the request, or
sometimes the target, so we cannot rely on it easily.

And as we always have the `target` key correctly filled, this commit
also adds the `trigger` key.

This also updates the `afterSwap.htmx` and `afterSettle.htmx` events to
be triggered for all newly added elements when `hx-swap` is
`beforebegin` and `beforeend`, for consistency with other modes.

Tests: many are added in `test/core/events.js` to check many cases.
Doc: `www/events.md` is updated accordingly.