Adding websocket retries
This patch adds retries to WebSockets when a connection fails or is closed.

It uses a randomized Exponential Backoff algorithm to try to mitigate large numbers of clients from flooding servers at once -- and allows only a single  timer to run, in case the browser triggers multiple "onclose" events in quick succession.