First attempt, not working.  DO NOT USE.
Separated out some functions to be used in common between WebSockets and SSE, but SSE is not working correctly.  Old content is removed, but the new content is not added to the DOM.

I've traced the errors down into the swapOuterHTML function, but am not sure what's causing the problem.

Checking in so that I can try something else on another branch.