Closing connections BEFORE removing nodes
In some tests, the previous code triggered an htmx:swapError because the node had already been removed from the DOM before the closeConnections() function was called.

I think it was a race condition that I didn't catch earlier.  Switching the ordering of these two functions makes sense, and it looks like it's addressed the error.

Sorry for the sloppy check-in.