Connection Pooling 2.0
This is a first draft of a better kind of connection pooling.  Since the number of connections is unlikely to grow very large, it seems easier and more efficient to just put them all into one array, and scan through it whenever we need.

This enables a new function `auditEventListeners` which verifies that the (probably < 4) remote connections are still supporting active DOM nodes, and saves us the extra work of walking an entire sub-tree hunting for them.

I think this same structure could save us even more work if applied to WebSockets, too.  But, that effort should wait until we've confirmed that we like this for SSE.