Adding limit: argument to hx-swap command
This adds a limit to the number of child elements that swap will add to a node.  If the limit is exceeded, then extra children are removed.

If we're adding elements to the END of the list, then extra nodes are removed from the BEGINNING.

Conversely, if we're adding to the BEGINNING, then we remove from the END.