Fix throttling to avoid delaying
This changes throttling so it doesn't wait to be fired when the event happens, _unless_ it has already been fired and a timeout is waiting to be executed to clean itself.

Please note this is a bit of a shoot in the dark as I'm editing this directly in GitHub and haven't tested it, and of course I'm not adding/fixing any automated tests. So it's just an idea. :P