BUGFIX: API call to shouldCancel
This addresses problems using the WebSocket extension to send messages to a websocket server.  We were calling `shouldCancel` with the incorrect API signature, causing it to return FALSE (as in, don't cancel the event) when it should have returned TRUE (as in, yes, cancel the event)

I don't know why this worked in my tests before, but the test case it behaving correctly now.