[RFC] Always use latest `href`/`hx-*` in `hx-boost`
hx-boost seems to capture the href/hx-get/etc of an element when the boosting logic is setup. 

But if the element changes its `href` since it was created, `hx-boost` will cause it to redirect to the wrong location. While opening that link in a new tab/window (bypassing hx-boost) will sent to the correct URL.

This proposed change would -- when performing an AJAX request -- revalidate that the element should be boosted and compute the verb/path at that time.

Thanks for your consideration! :)