Add documentation about HTTP caching
Everything works out of the box as expected. Users that generate
different responses on the same URL based on some header need to be
mindful of the implications (for example using the Vary response
header). This is mentioned because it's a common pattern to use
HX-Request header to render only an HTML fragment.