Escape curly braces in the example (#1175)
Docs on the website are currently rendering incorrectly because of faulty escaping of curly braces (line 80 is rendered has `<p> and  and </> htmx - high power tools for html and </p>`). Unfortunately, this fix breaks the rendering of the file on github (there are extra `{% raw %}` directives), but I think it's better than the other way around.