Fix normalizeURL implementation
The URL constructor does not work with relative URLs unless you
explicitly specify a base URL. HTMX 1.0 wrapped this function in a
try/catch block for IE11 support, but that try/catch was also catching a
lot of stuff that modern browsers should have been normalizing, because
relative URLs would just throw an excpetion and return themselves.

Rather than add back the try/catch, I just check if it's a relative URL
and add the window origin to the URL if so. That way the function will
continue to throw exceptions for malformatted URLs, which seems
appropriate.

This change fixes the test suite so the HTMX 2.0 tests pass.