Add date and author to essay template
Update the static site generator to include the date and author
automatically as part of the essays. I made a couple changes to the
underlying static structure to make this a little more seamless,
including:

* New "essay.html" template for posts in `/essays`
* Added an author taxonomy (can eventually be author pages)
* Added Carson as the author of all the existing esssays
* Cleaned up manual date entires, where they existed

All this will make it easier to publish guest essays and sort the essays
(chronologically, by topic, or by author) as the essay base expands.