Document how to use event delegation within hx-trigger for dynamically added elements
This PR updates the `hx-trigger` attribute documentation to explain how to use event delegation for dynamically added elements.

The new information helps users understand how to:
1. Use event delegation for dynamically added elements
2. Avoid syntax errors when using selectors with spaces

Previously, the documentation didn't cover how to handle event delegation for elements added to the DOM after the initial page load. I struggled with htmx errors due to white space issues when referencing the dynamic element. So I thought it's worth documenting.

The update adds information on using curly braces `{}` to the `from:` modifier of `hx-trigger`. 

This documentation update corresponds to the implementation in PR #1913, which introduced this functionality to HTMX.