Update CONTRIBUTING.md  - fixing a link (#2795)
* Update CONTRIBUTING.md  - fixing a link, wording

Excited to see this unframework framework as a long time (since 1998 coder) HTML appreciator. 

I wanted to learn about extensions but noticed a broken link that others might stumble upon. 

Consider rephrasing a question to invite a person to contribute their issue. (this can be ignored if you like).

* Update CONTRIBUTING.md

Removed change to grammar/style of writing back to original.
Providing a more updated reference to extensions with greater detail.