fix vertical alignment of github stars button in all states
Without uBlock, the button's alt-text would be 1px lower than the rest
of the nav items, and inconsistently capitalized.

With uBlock, the button would be slightly too high up.

This is on Firefox.

The solution is to apply styles on the `span` that GitHub's JS inserts.
However, using margin will make the entire nav jump during pageload, so
position: relative seems more robust there.