Correct the integrity hash of the scripts 'Integrity' key for htmx-ext-ws@2.0.2
Looks like there is a % typo on the end of the integrity field.

This is likely because a command was used like this:

➜ cat htmx-ext-ws@2.0.2 | openssl dgst -sha384 -binary | openssl base64 -A
0RHpPqZ4QLJXG/ALtieEL/G+NuAI98LYSkT9s6FgciveUhAdo/6wab5NU1tm2Bxs%

Note that it adds a % on the end, this is the shell end character being added.

To prevent this, you can instruct your shell to add a newline after the command's output.
The easy to do this is to add an echo command after the pipeline.

➜ cat htmx-ext-ws@2.0.2 | openssl dgst -sha384 -binary | openssl base64 -A; echo

0RHpPqZ4QLJXG/ALtieEL/G+NuAI98LYSkT9s6FgciveUhAdo/6wab5NU1tm2Bxs

This only addresses the documentation, I don't know if you have tooling that generates
the value in the docs.

Thank you.