Rework multiple spec and multiple targets for delayed modifier
If multiple nodes are targetted by one trigger specification, they share a delay.
This is useful e.g. for active searches with multiple participating
fields, so that the search triggers only when search modifiers have been
processed.

If a node has multiple triggers with the delay modifier, they act independently.
This corresponds to the logic from #2026.